Add TransmitterBase
